+++ crates/plugins/mediator-coordination/src/web/handler.rs
@@ -8,7 +8,7 @@ use hyper::{header::CONTENT_TYPE, StatusCode};
 use std::sync::Arc;
 
 use crate::{
-    constant::{DIDCOMM_ENCRYPTED_MIME_TYPE, KEYLIST_UPDATE_2_0, MEDIATE_REQUEST_2_0},
+    constant::{DIDCOMM_ENCRYPTED_MIME_TYPE, KEYLIST_UPDATE_2_0, KEYLIST_QUERY_2_0},
     web::{self, error::MediationError, AppState},
 };
 
@@ -17,7 +17,6 @@ pub async fn process_didcomm_message(
     State(state): State<Arc<AppState>>,
     Extension(message): Extension<Message>,
 ) -> Response {
-    // handle mediation request
     let delegate_response = match message.type_.as_str() {
         KEYLIST_UPDATE_2_0 => {
             web::coord::handler::stateful::process_plain_keylist_update_message(
@@ -25,9 +24,13 @@ pub async fn process_didcomm_message(
                 message,
             )
             .await
-        }
-        MEDIATE_REQUEST_2_0 => {
-            web::coord::handler::stateful::process_mediate_request(&state, &message).await
+        },
+        KEYLIST_QUERY_2_0 => {
+            web::coord::handler::stateful::process_plain_keylist_query_message(
+                Arc::clone(&state),
+                message,
+            )
+            .await
         }
         _ => {
             let response = (
